API reference

Types

Functions

BinaryBuilder.autobuild — Method

autobuild(dir::AbstractString, src_name::AbstractString,
          src_version::VersionNumber, sources::Vector,
          script::AbstractString, platforms::Vector,
          products::Vector, dependencies::Vector;
          verbose = false, debug = false,
          skip_audit = false, ignore_audit_errors = true,
          autofix = true, code_dir = nothing,
          meta_json_file = nothing, require_license = true, kwargs...)

Runs the boiler plate code to download, build, and package a source package for a list of platforms. This method takes a veritable truckload of arguments, here are the relevant actors, broken down in brief:

dir: the root of the build; products will be placed within dir/products, and mountpoints will be placed within dir/build/.

src_name: the name of the source package being built and will set the name of the built tarballs.

src_version: the version of the source package.

platforms: a list of platforms to build for.

sources: a vector of all sources to download and unpack before building begins, as AbstractSources.

script: a string representing a shell script to run as the build.

products: the list of Products which shall be built.

dependencies: a vector of JLL dependency packages as AbstractDependency that should be installed before building begins.

verbose: Enable verbose mode. What did you expect?

debug: cause a failed build to drop into an interactive shell so that the build can be inspected easily.

skip_audit: disable the typical audit that occurs at the end of a build.

ignore_audit_errors: do not kill a build even if a problem is found.

autofix: give BinaryBuilder permission to automatically fix issues it finds during audit passes. Highly recommended.

code_dir: sets where autogenerated JLL packages will be put.

require_license enables a special audit pass that requires licenses to be installed by all packages.

lazy_artifacts sets whether the artifacts should be lazy.

meta_json_stream: If this is set to an IOStream, do not actually build, just output a JSON representation of all the metadata about this build to the stream.

BinaryBuilder.build_tarballs — Method

build_tarballs(ARGS, src_name, src_version, sources, script, platforms,
               products, dependencies; kwargs...)

This should be the top-level function called from a build_tarballs.jl file. It takes in the information baked into a build_tarballs.jl file such as the sources to download, the products to build, etc... and will automatically download, build and package the tarballs, generating a build.jl file when appropriate. Note that ARGS should be the top-level Julia ARGS command- line arguments object. This function does some rudimentary parsing of the ARGS, call it with --help in the ARGS to see what it can do.

The kwargs are passed on to autobuild, see there for a list of supported ones. In addition, the keyword argument init_block may be set to a string containing Julia code; if present, this code will be inserted into the initialization path of the generated JLL package. This can for example be used to invoke an initialization API of a shared library.

Note

The init_block keyword argument is experimental and may be removed in a future version of this package. Please use it sparingly.

BinaryBuilder.get_compilers_versions — Method

get_compilers_versions(; compilers = [:c])

Return the script string that is used to print the versions of the given compilers.

BinaryBuilder.versioninfo — Method

versioninfo()

Helper function to print out some debugging information