API reference

Types

BinaryBuilder.BuildDependency — Type

BuildDependency(dep::Union{PackageSpec,String})

Define a binary dependency that is necessary only to build the package. The argument can be either a string with the name of the JLL package or a Pkg.PackageSpec.

BinaryBuilder.Dependency — Type

Dependency(dep::Union{PackageSpec,String})

Define a binary dependency that is necessary to build the package and load the generated JLL package. The argument can be either a string with the name of the JLL package or a Pkg.PackageSpec.

BinaryBuilder.DirectorySource — Type

DirectorySource(path::String; unpack_target::String = "")

Specify a local directory to mount from path.

The content of the directory will be mounted in ${WORKSPACE}/srcdir, or in its subdirectory pointed to by the optional keyword unpack_target, if provided.

BinaryBuilder.ExecutableProduct — Type

An ExecutableProduct is a Product that represents an executable file.

On all platforms, an ExecutableProduct checks for existence of the file. On non-Windows platforms, it will check for the executable bit being set. On Windows platforms, it will check that the file ends with ".exe", (adding it on automatically, if it is not already present).

BinaryBuilder.FileProduct — Type

FileProduct(path::AbstractString, varname::Symbol, dir_path = nothing)

Declares a FileProduct that points to a file located relative to the root of a Prefix, must simply exist to be satisfied.

BinaryBuilder.FileSource — Type

FileSource(url::String, hash::String; unpack_target::String = "")

Specify a remote file to be downloaded from the Internet from url. hash is the 64-character SHA256 checksum of the file.

If the file is an archive in one of the supported archive formats (e.g., TAR or ZIP balls), it will be automatically unpacked to ${WORKSPACE}/srcdir, or in its subdirectory pointed to by the optional keyword unpack_target, if provided.

BinaryBuilder.GitSource — Type

GitSource(url::String, hash::String; unpack_target::String = "")

Specify a remote Git repository to clone form url. hash is the 40-character SHA1 revision to checkout after cloning.

The repository will be cloned in ${WORKSPACE}/srcdir, or in its subdirectory pointed to by the optional keyword unpack_target, if provided.

BinaryBuilder.LibraryProduct — Type

A LibraryProduct is a special kind of Product that not only needs to exist, but needs to be dlopen()'able. You must know which directory the library will be installed to, and its name, e.g. to build a LibraryProduct that refers to "/lib/libnettle.so", the "directory" would be "/lib", and the "libname" would be "libnettle". Note that a LibraryProduct can support multiple libnames, as some software projects change the libname based on the build configuration.

BinaryBuilder.OutputCollector — Type

OutputCollector

A run() wrapper class that captures subprocess stdout and stderr streams independently, resynthesizing and colorizing the streams appropriately.

BinaryBuilder.OutputCollector — Method

OutputCollector(cmd::AbstractCmd; verbose::Bool = false)

Run cmd, and collect the output such that stdout and stderr are captured independently, but with the time of each line recorded such that they can be stored/analyzed independently, but replayed synchronously.

BinaryBuilder.Product — Type

A Product is an expected result after building or installation of a package.

Examples of Products include LibraryProduct, ExecutableProduct and FileProduct. All Product types must define the following minimum set of functionality:

locate(::Product): given a Product, locate it within the wrapped Prefix returning its location as a string

satisfied(::Product): given a Product, determine whether it has been successfully satisfied (e.g. it is locateable and it passes all callbacks)

variable_name(::Product): return the variable name assigned to a Product

repr(::Product): Return a representation of this Product, useful for auto-generating source code that constructs Products, if that's your thing.

BinaryBuilder.AbstractDependency — Type

An AbstractDependency is a binary dependency of the JLL package. Dependencies are installed to ${prefix} in the build environment.

Concrete subtypes of AbstractDependency are

Dependency: a JLL package that is necessary for to build the package and to load the generated JLL package.

BuildDependency: a JLL package that is necessary only t obuild the package. This will not be a dependency of the generated JLL package.

BinaryBuilder.AbstractSource — Type

An AbstractSource is something used as source to build the package. Sources are installed to ${WORKSPACE}/srcdir in the build environment.

Concrete subtypes of AbstractSource are:

FileSource: a remote file to download from the Internet;

GitSource: a remote Git repository to clone;

DirectorySource: a local directory to mount.

BinaryBuilder.DockerRunner — Type

DockerRunner

Use docker as an execution engine; a reasonable backup for platforms that do not have user namespaces (e.g. MacOS, Windows).

BinaryBuilder.LineStream — Method

LineStream(pipe::Pipe)

Given a Pipe that has been initialized by spawn(), create an async Task to read in lines as they come in and annotate the time the line was captured for later replay/merging with other simultaneously captured streams.

BinaryBuilder.UserNSRunner — Type

UserNSRunner

A UserNSRunner represents an "execution context", an object that bundles all necessary information to run commands within the container that contains our crossbuild environment. Use run() to actually run commands within the UserNSRunner, and runshell() as a quick way to get an interactive shell within the crossbuild environment.

BinaryBuilder.WizardState — Type

WizardState

Building large dependencies can take a lot of time. This state object captures all relevant state of this function. It can be passed back to the function to resume where we left off. This can aid debugging when code changes are necessary. It also holds all necessary metadata such as input/output streams.

Functions

Base.merge — Method

merge(collector::OutputCollector; colored::Bool = false)

Merge the stdout and stderr streams of the OutputCollector on a per-line basis, returning a single string containing all collected lines, interleaved by capture time. If colored is set to true, embeds terminal color codes to print stderr in red.

BinaryBuilder.audit — Function

audit(prefix::Prefix; platform::Platform = platform_key_abi();
                      verbose::Bool = false,
                      silent::Bool = false,
                      autofix::Bool = false,
                      require_license::Bool = true)

Audits a prefix to attempt to find deployability issues with the binary objects that have been installed within. This auditing will check for relocatability issues such as dependencies on libraries outside of the current prefix, usage of advanced instruction sets such as AVX2 that may not be usable on many platforms, linkage against newer glibc symbols, etc...

This method is still a work in progress, only some of the above list is actually implemented, be sure to actually inspect Auditor.jl to see what is and is not currently in the realm of fantasy.

BinaryBuilder.autobuild — Method

autobuild(dir::AbstractString, src_name::AbstractString,
          src_version::VersionNumber, sources::Vector,
          script::AbstractString, platforms::Vector,
          products::Vector, dependencies::Vector;
          verbose = false, debug = false,
          skip_audit = false, ignore_audit_errors = true,
          autofix = true, code_dir = nothing,
          meta_json_file = nothing, require_license = true, kwargs...)

Runs the boiler plate code to download, build, and package a source package for a list of platforms. This method takes a veritable truckload of arguments, here are the relevant actors, broken down in brief:

dir: the root of the build; products will be placed within dir/products, and mountpoints will be placed within dir/build/.

src_name: the name of the source package being built and will set the name of the built tarballs.

src_version: the version of the source package.

platforms: a list of platforms to build for.

sources: a vector of all sources to download and unpack before building begins, as AbstractSources.

script: a string representing a shell script to run as the build.

products: the list of Products which shall be built.

dependencies: a vector of JLL dependency packages as AbstractDependency that should be installed before building begins.

verbose: Enable verbose mode. What did you expect?

debug: cause a failed build to drop into an interactive shell so that the build can be inspected easily.

skip_audit: disable the typical audit that occurs at the end of a build.

ignore_audit_errors: do not kill a build even if a problem is found.

autofix: give BinaryBuilder permission to automatically fix issues it finds during audit passes. Highly recommended.

code_dir: sets where autogenerated JLL packages will be put.

require_license enables a special audit pass that requires licenses to be installed by all packages.

lazy_artifacts sets whether the artifacts should be lazy.

meta_json_stream: If this is set to an IOStream, do not actually build, just output a JSON representation of all the metadata about this build to the stream.

BinaryBuilder.bindir — Method

bindir(prefix::Prefix)

Returns the binary directory for the given prefix.

BinaryBuilder.build_tarballs — Method

build_tarballs(ARGS, src_name, src_version, sources, script, platforms,
               products, dependencies; kwargs...)

This should be the top-level function called from a build_tarballs.jl file. It takes in the information baked into a build_tarballs.jl file such as the sources to download, the products to build, etc... and will automatically download, build and package the tarballs, generating a build.jl file when appropriate. Note that ARGS should be the top-level Julia ARGS command- line arguments object. This function does some rudimentary parsing of the ARGS, call it with --help in the ARGS to see what it can do.

BinaryBuilder.collapse_symlinks — Method

collapse_symlinks(files::Vector{String})

Given a list of files, prune those that are symlinks pointing to other files within the list.

BinaryBuilder.collect_files — Function

collect_files(path::AbstractString, predicate::Function = f -> true)

Find all files that satisfy predicate() when the full path to that file is passed in, returning the list of file paths.

BinaryBuilder.collect_stderr — Method

collect_stderr(collector::OutputCollector)

Returns all stderr lines collected by this collector so far.

BinaryBuilder.collect_stdout — Method

collect_stdout(collector::OutputCollector)

Returns all stdout lines collected by this collector so far.

BinaryBuilder.expand_cxxstring_abis — Method

expand_cxxstring_abis(p::Platform)

Given a Platform, returns an array of Platforms with a spread of identical entries with the exception of the cxxstring_abi member of the CompilerABI struct within the Platform. This is used to take, for example, a list of supported platforms and expand them to include multiple GCC versions for the purposes of ABI matching. If the given Platform already specifies a GCC version (as opposed to nothing) only that Platform is returned.

BinaryBuilder.expand_gfortran_versions — Method

expand_gfortran_versions(p::Platform)

Given a Platform, returns an array of Platforms with a spread of identical entries with the exception of the gcc_version member of the CompilerABI struct within the Platform. This is used to take, for example, a list of supported platforms and expand them to include multiple GCC versions for the purposes of ABI matching. If the given Platform already specifies a GCC version (as opposed to nothing) only that Platform is returned.

BinaryBuilder.includedir — Method

includedir(prefix::Prefix)

Returns the include directory for the given prefix

BinaryBuilder.libdirs — Function

libdirs(prefix::Prefix, platform = platform_key_abi())

Returns the library directories for the given prefix (note that this differs between unix systems and windows systems, and between 32- and 64-bit systems).

BinaryBuilder.locate — Method

locate(fp::ExecutableProduct, prefix::Prefix;
       platform::Platform = platform_key_abi(),
       verbose::Bool = false,
       isolate::Bool = false)

If the given executable file exists and is executable, return its path.

On all platforms, an ExecutableProduct checks for existence of the file. On non-Windows platforms, it will check for the executable bit being set. On Windows platforms, it will check that the file ends with ".exe", (adding it on automatically, if it is not already present).

BinaryBuilder.locate — Method

locate(fp::FileProduct, prefix::Prefix;
       platform::Platform = platform_key_abi(),
       verbose::Bool = false,
       isolate::Bool = false)

If the given file exists, return its path. The platform and isolate arguments are is ignored here, but included for uniformity. For ease of use, we support a limited number of custom variable expansions such as ${target}, and ${nbits}, so that the detection of files within target-specific folders named things like /lib32/i686-linux-musl is simpler.

BinaryBuilder.locate — Method

locate(lp::LibraryProduct, prefix::Prefix;
       verbose::Bool = false,
       platform::Platform = platform_key_abi())

If the given library exists (under any reasonable name) and is dlopen()able, (assuming it was built for the current platform) return its location. Note that the dlopen() test is only run if the current platform matches the given platform keyword argument, as cross-compiled libraries cannot be dlopen()ed on foreign platforms.

BinaryBuilder.logdir — Method

logdir(prefix::Prefix)

Returns the logs directory for the given prefix.

BinaryBuilder.satisfied — Method

satisfied(p::Product;
          platform::Platform = platform_key_abi(),
          verbose::Bool = false,
          isolate::Bool = false)

Given a Product, return true if that Product is satisfied, e.g. whether a file exists that matches all criteria setup for that Product. If isolate is set to true, will isolate all checks from the main Julia process in the event that dlopen()'ing a library might cause issues.

BinaryBuilder.supported_platforms — Method

supported_platforms(;exclude::Union{Vector{<:Platform},Function}=x->false)

Return the list of supported platforms as an array of Platforms. These are the platforms we officially support building for, if you see a mapping in get_shard_hash() that isn't represented here, it's probably because that platform is still considered "in beta".

Platforms can be excluded from the list by specifying an array of platforms to exclude i.e. supported_platforms(exclude=[Windows(:i686),Windows(:x86_64)]) or a function that returns true for exclusions i.e.

islin(x) = typeof(x) == Linux
supported_platforms(exclude=islin)

BinaryBuilder.tail — Method

tail(collector::OutputCollector; len::Int = 100, colored::Bool = false)

Write out the last len lines, optionally writing colored lines.

BinaryBuilder.tee — Method

tee(c::OutputCollector; colored::Bool = false, stream::IO = stdout)

Spawn a background task to incrementally output lines from collector to the standard output, optionally colored.

BinaryBuilder.temp_prefix — Method

temp_prefix(func::Function)

Create a temporary prefix, passing the prefix into the user-defined function so that build/packaging operations can occur within the temporary prefix, which is then cleaned up after all operations are finished. If the path provided exists already, it will be deleted.

Usage example:

out_path = abspath("./libfoo")
temp_prefix() do p
    # <insert build steps here>

    # tarball up the built package
    tarball_path, tarball_hash = package(p, out_path)
end

BinaryBuilder.variable_name — Method

variable_name(p::Product)

Return the variable name associated with this Product as a string

Pkg.BinaryPlatforms.detect_cxxstring_abi — Method

detect_cxxstring_abi(oh::ObjectHandle, platform::Platform)

Given an ObjectFile, examine its symbols to discover which (if any) C++11 std::string ABI it's using. We do this by scanning the list of exported symbols, triggering off of instances of St7__cxx11 or _ZNSs to give evidence toward a constraint on cxx11, cxx03 or neither.

Pkg.BinaryPlatforms.detect_libgfortran_version — Method

detect_libgfortran_version(oh::ObjectHandle, platform::Platform)

Given an ObjectFile, examine its dynamic linkage to discover which (if any) libgfortran it's linked against. The major SOVERSION will determine which GCC version we're restricted to.

Pkg.BinaryPlatforms.detect_libstdcxx_version — Method

detect_libstdcxx_version(oh::ObjectHandle, platform::Platform)

Given an ObjectFile, examine its dynamic linkage to discover which (if any) libgfortran it's linked against. The major SOVERSION will determine which GCC version we're restricted to.

Base.wait — Method

wait(collector::OutputCollector)

Wait for the command and all line streams within an OutputCollector to finish their respective tasks and be ready for full merging. Return the success of the underlying process. Prints out the last 10 lines of the process if it does not complete successfully unless the OutputCollector was created as verbose.

BinaryBuilder.abi_agnostic — Method

abi_agnostic(p::Platform)

Strip out the CompilerABI portion of a Platform, making it "ABI agnostic".

BinaryBuilder.alive — Method

alive(s::LineStream)

Returns true if the task owned by this LineStream is still processing output from an underlying Pipe.

BinaryBuilder.analyze_instruction_set — Method

analyze_instruction_set(oh::ObjectHandle, platform::Platform; verbose::Bool = false)

Analyze the instructions within the binary located at the given path for which minimum instruction set it requires, taking note of groups of instruction sets used such as avx, sse4.2, i486, etc....

Some binary files (such as libopenblas) contain multiple versions of functions, internally determining which version to call by using the cpuid instruction to determine processor support. In an effort to detect this, we make note of any usage of the cpuid instruction, disabling our minimum instruction set calculations if such an instruction is found, and notifying the user of this if verbose is set to true.

Note that this function only really makes sense for x86/x64 binaries. Don't run this on armv7l, aarch64, ppc64le etc... binaries and expect it to work.

BinaryBuilder.artifact_name — Method

artifact_name(cs::CompilerShard)

Return the bound artifact name for a particular shard.

BinaryBuilder.canonicalize_file_url — Method

Canonicalize URL to a file within a GitHub repo

BinaryBuilder.canonicalize_source_url — Method

Canonicalize a GitHub repository URL

BinaryBuilder.change_script! — Method

Change the script. This will invalidate all platforms to make sure we later
verify that they still build with the new script.

BinaryBuilder.check_license — Function

check_license(prefix, src_name; verbose::Bool = false,, silent::Bool = false)

Check that there are license files for the project called src_name in the prefix.

BinaryBuilder.choose_shards — Method

choose_shards(p::Platform; rootfs_build, ps_build, GCC_builds,
                           LLVM_builds, archive_type)

This method chooses, given a Platform, which shards to download, extract and mount, returning a list of CompilerShard objects. At the moment, this always consists of four shards, but that may not always be the case.

BinaryBuilder.chown_cleanup — Method

chown_cleanup(dr::DockerRunner)

On Linux, the user id inside of the docker container doesn't correspond to ours on the outside, so permissions get all kinds of screwed up. To fix this, we have to chown -R $(id -u):$(id -g) $prefix, which really sucks, but is still better than nothing. This is why we prefer the UserNSRunner on Linux.

BinaryBuilder.clone — Method

clone(url::String, source_path::String)

Clone a git repository hosted at url into source_path, with a progress bar displayed to stdout.

BinaryBuilder.create_and_bind_mutable_artifact! — Method

create_and_bind_mutable_artifact!(f::Function, art_name::String)

Create (and bind) an artifact to MutableArtifacts.toml in one fell swoop. Used in things like .squashfs UID remapping and BB wizard state serialization.

BinaryBuilder.download_all_artifacts — Method

download_all_shards(; verbose::Bool=false)

Helper function to download all shards/helper binaries so that no matter what happens, you don't need an internet connection to build your precious, precious binaries.

BinaryBuilder.download_source — Function

download_source(source::AbstractSource; verbose::Bool = false)

Download the given source. All downloads are cached within the BinaryBuilder downloads storage directory.

BinaryBuilder.download_source — Method

download_source(state::WizardState)

Ask the user where the source code is coming from, then download and record the relevant parameters, returning the source url, the local path it is stored at after download, and a hash identifying the version of the code. In the case of a git source URL, the hash will be a git treeish identifying the exact commit used to build the code, in the case of a tarball, it is the sha256 hash of the tarball itself.

BinaryBuilder.edit_script — Method

edit_script(state::WizardState, script::AbstractString)

For consistency (and security), use the sandbox for editing a script, launching vi within an interactive session to edit a buildscript.

BinaryBuilder.filter_object_files — Method

filter_object_files(files)

Given a list of files, filter out any that cannot be opened by readmeta() from ObjectFile.

BinaryBuilder.gcc_version — Method

gcc_version(cabi::CompilerABI, GCC_builds::Vector{GCCBuild})

Returns the closest matching GCC version number for the given CompilerABI representing a particular platofrm, from the given set of options. If no match is found, returns an empty list. This method assumes that cabi represents a platform that binaries will be run on, and thus versions are always rounded down; e.g. if the platform supports a libstdc++ version that corresponds to GCC 5.1.0, but the only GCC versions available to be picked from are 4.8.5 and 5.2.0, it will return 4.8.5, as binaries compiled with that version will run on this platform, whereas binaries compiled with 5.2.0 may not.

BinaryBuilder.generate_compiler_wrappers! — Method

generate_compiler_wrappers!(platform::Platform; bin_path::AbstractString,
                            host_platform::Platform = Linux(:x86_64; libc=:musl),
                            rust_platform::Platform = Linux(:x86_64; libc=:glibc),
                            compilers::Vector{Symbol} = [:c],
                            allow_unsafe_flags::Bool = false)

We generate a set of compiler wrapper scripts within our build environment to force all build systems to honor the necessary sets of compiler flags to build for our systems. Note that while platform_envs() sets many environment variables, those values are intended to be optional/overridable. These values, while still overridable by directly invoking a compiler binary directly (e.g. /opt/{target}/bin/{target}-gcc), are much more difficult to override, as the flags embedded in these wrappers are absolutely necessary, and even simple programs will not compile without them.

BinaryBuilder.generate_per_uid_squashfs — Function

generate_per_uid_squashfs(cs, new_uid = getuid())

In order for the sandbox to work well, we need to have the uids of the squashfs images match the uid of the current unpriviledged user. Unfortunately there is no mount-time option to do this for us. Fortunately, squashfs is simple enough that if the ID table is uncompressed, we can just manually patch the uids to be what we need. This function performs this operation, by rewriting all UIDs and GIDs to the given new_uid (which defaults to the current user's UID)..

BinaryBuilder.get_compilers_versions — Method

get_compilers_versions(; compilers = [:c])

Return the script string that is used to print the versions of the given compilers.

BinaryBuilder.get_mutable_artifact_path — Method

get_mutable_artifact_path(art_name::String)

Convenience wrapper to get an artifact bound withing MutableArtifacts.toml. Returns nothing if artifact not bound yet.

BinaryBuilder.getgid — Method

getgid()

Wrapper around libc's getgid() function

BinaryBuilder.getuid — Method

getuid()

Wrapper around libc's getuid() function

BinaryBuilder.import_docker_image — Method

import_docker_image(rootfs::CompilerShard; verbose::Bool = false)

Checks to see if the given rootfs has been imported into docker yet; if it hasn't, then do so so that we can run things like:

docker run -ti binarybuilder_rootfs:v2018.08.27 /bin/bash

Which, after all, is the foundation upon which this whole doodad is built.

BinaryBuilder.instruction_mnemonics — Method

instruction_mnemonics(path::AbstractString, platform::Platform)

Dump a binary object with objdump, returning a list of instruction mnemonics for further analysis with analyze_instruction_set().

Note that this function only really makes sense for x86/x64 binaries. Don't run this on armv7l, aarch64, ppc64le etc... binaries and expect it to work.

This function returns the list of mnemonics as well as the counts of each, binned by the mapping defined within instruction_categories.

BinaryBuilder.interactive_build — Method

interactive_build(state::WizardState, prefix::Prefix,
                  ur::Runner, build_path::AbstractString)

Runs the interactive shell for building, then captures bash history to save
reproducible steps for building this source. Shared between steps 3 and 5

BinaryBuilder.is_ecryptfs — Method

is_ecryptfs(path::AbstractString; verbose::Bool=false)

Checks to see if the given path (or any parent directory) is placed upon an ecryptfs mount. This is known not to work on current kernels, see this bug for more details: https://bugzilla.kernel.org/show_bug.cgi?id=197603

This method returns whether it is encrypted or not, and what mountpoint it used to make that decision.

BinaryBuilder.is_for_platform — Method

is_for_platform(h::ObjectHandle, platform::Platform)

Returns true if the given ObjectHandle refers to an object of the given platform; E.g. if the given platform is for AArch64 Linux, then h must be an ELFHandle with h.header.e_machine set to ELF.EM_AARCH64.

In particular, this method and platform_for_object() both exist because the latter is not smart enough to deal with :glibc and :musl yet.

BinaryBuilder.is_mounted — Method

is_mounted(cs::CompilerShard, build_prefix::String)

Return true if the given shard is mounted. Uses run() so will error out if something goes awry.

BinaryBuilder.is_troublesome_library_link — Method

is_troublesome_library_link(libname::AbstractString, platform::Platform)

Return true if depending on libname is known to cause problems at runtime, false otherwise.

BinaryBuilder.macos_sdk_already_installed — Method

macos_sdk_already_installed()

Returns true if any piece of the MacOS SDK is already installed.

BinaryBuilder.map_target — Method

map_target(cs::CompilerShard)

Return the location this compiler shard should be mounted at. We basically analyze the name and platform of this shard and return a path based on that.

BinaryBuilder.match_files — Method

match_files(state::WizardState, prefix::Prefix,
            platform::Platform, files::Vector; silent::Bool = false)

Inspects all binary files within a prefix, matching them with a given list of files, complaining if there are any files that are not properly matched and returning the set of normalized names that were not matched, or an empty set if all names were properly matched.

BinaryBuilder.minimum_instruction_set — Method

minimum_instruction_set(counts::Dict, is_64bit::Bool)

This function returns the minimum instruction set required, depending on whether the object file being pointed to is a 32-bit or 64-bit one:

For 32-bit object files, this returns one of [:pentium4, :prescott]

For 64-bit object files, this returns one of [:core2, :sandybridge, :haswell]

BinaryBuilder.mount — Method

mount(cs::CompilerShard, build_prefix::String)

Mount a compiler shard, if possible. Uses run() so will error out if something goes awry. Note that this function only does something when using a .squashfs shard, with a UserNS or Docker runner, on Linux. All other combinations of shard archive type, runner and platform result in a no-op from this function.

BinaryBuilder.normalize_name — Method

normalize_name(file::AbstractString)

Given a filename, normalize it, stripping out extensions. E.g. the file path "foo/libfoo.tar.gz" would get mapped to "libfoo".

BinaryBuilder.pick_preferred_platform — Method

Pick the first platform for use to run on. We prefer Linux x86_64 because that's generally the host platform, so it's usually easiest. After that we go by the following preferences:

OS (in order): Linux, Windows, OSX

Architecture: x86_64, i686, aarch64, powerpc64le, armv7l

The first remaining after this selection

BinaryBuilder.platform_envs — Method

platform_envs(platform::Platform)

Given a platform, generate a Dict mapping representing all the environment variables to be set within the build environment to force compiles toward the defined target architecture. Examples of things set are PATH, CC, RANLIB, as well as nonstandard things like target.

BinaryBuilder.platform_for_object — Method

platform_for_object(oh::ObjectHandle)

Returns the platform the given ObjectHandle should run on. E.g. if the given ObjectHandle is an x86_64 Linux ELF object, this function will return Linux(:x86_64). This function does not yet distinguish between different libc's such as :glibc and :musl.

BinaryBuilder.preferred_cxxstring_abi — Method

preferred_cxxstring_abi(platform::Platform, shard::CompilerShard;
                        gcc_builds::Vector{GCCBuild} = available_gcc_builds)

Return the C++ string ABI preferred by the given platform or GCCBootstrap shard.

BinaryBuilder.preferred_libgfortran_version — Method

preferred_libgfortran_version(platform::Platform, shard::CompilerShard;
                              gcc_builds::Vector{GCCBuild} = available_gcc_builds)

Return the libgfortran version preferred by the given platform or GCCBootstrap shard.

BinaryBuilder.print_autoconf_hint — Method

print_autoconf_hint(state::WizardState)

Print a hint for projets that use autoconf to have a good ./configure line.

BinaryBuilder.provide_hints — Method

provide_hints(state::WizardState, path::AbstractString)

Given an unpacked source directory, provide hints on how a user might go about building the binary bounty they so richly desire.

BinaryBuilder.readuntil_many — Method

readuntil_many(s::IO, delims)

Given a collection of delimiter characters, read from s until one of those delimiters is reached, or we reach the end of s.

BinaryBuilder.runshell — Function

runshell(platform::Platform = platform_key_abi())

Launch an interactive shell session within the user namespace, with environment setup to target the given platform.

BinaryBuilder.setup_dependencies — Method

setup_dependencies(prefix::Prefix, dependencies::Vector{PackageSpec})

Given a list of JLL package specifiers, install their artifacts into the build prefix. The artifacts are installed into the global artifact store, then copied into a temporary location, then finally symlinked into the build prefix. This allows us to (a) save download bandwidth by not downloading the same artifacts over and over again, (b) maintain separation in the event of catastrophic containment failure, avoiding hosing the main system if a build script decides to try to modify the dependent artifact files, and (c) keeping a record of what files are a part of dependencies as opposed to the package being built, in the form of symlinks to a specific artifacts directory.

BinaryBuilder.setup_workspace — Method

setup_workspace(build_path::String, sources::Vector{SetupSource};
                verbose::Bool = false)

Sets up a workspace within build_path, creating the directory structure needed by further steps, unpacking the source within build_path, and defining the environment variables that will be defined within the sandbox environment.

This method returns the Prefix to install things into, and the runner that can be used to launch commands within this workspace.

BinaryBuilder.shard_mappings — Method

shard_mappings(shards::Vector{CompilerShard})

Return the default mappings for a set of compiler shards

BinaryBuilder.shard_path — Method

shard_path(cs::CompilerShard)

Return the path to this shard on-disk; for unpacked shards, this is a directory. For squashfs shards, this is a file. This will not cause a shard to be downloaded.

BinaryBuilder.step1 — Method

step1(state::WizardState)

It all starts with a single step, the unabashed ambition to leave your current stability and engage with the universe on a quest to create something new, beautiful and unforseen. It all ends with compiler errors.

This step selects the relevant platform(s) for the built binaries.

BinaryBuilder.step2 — Method

step2(state::WizardState)

This step obtains the source code to be built and required binary dependencies.

BinaryBuilder.step34 — Method

step34(state::WizardState)

Starts initial build for Linux x86_64, which is our initial test target platform. Sources that build properly for this platform continue on to attempt builds for more complex platforms.

BinaryBuilder.step3_audit — Method

step3_audit(state::WizardState, platform::Platform, prefix::Prefix)

Audit the prefix.

BinaryBuilder.step3_interactive — Method

step3_interactive(state::WizardState, prefix::Prefix, platform::Platform,
                  ur::Runner, build_path::AbstractString)

The interactive portion of step3, moving on to either rebuild with an edited script or proceed to step 4.

BinaryBuilder.step3_retry — Method

step3_retry(state::WizardState)

Rebuilds the initial Linux x86_64 build after things like editing the script file manually, etc...

BinaryBuilder.step4 — Method

step4(state::WizardState, ur::Runner, platform::Platform,
      build_path::AbstractString, prefix::Prefix)

The fourth step selects build products after the first build is done

BinaryBuilder.symlink_soname_lib — Method

symlink_soname_lib(path::AbstractString)

We require that all shared libraries are accessible on disk through their SONAME (if it exists). While this is almost always true in practice, it doesn't hurt to make doubly sure.

BinaryBuilder.translate_symlinks — Method

translate_symlinks(root::AbstractString; verbose::Bool=false)

Walks through the root directory given within root, finding all symlinks that point to an absolute path within root, and rewriting them to be a relative symlink instead, increasing relocatability.

BinaryBuilder.uname — Method

uname()

On Linux systems, return the strings returned by the uname() function in libc

BinaryBuilder.unmount — Method

unmount(cs::CompilerShard, build_prefix::String)

Unmount a compiler shard from a given build prefix, if possible. Uses run() so will error out if something goes awry. Note that this function only does something when using a squashfs shard on Linux. All other combinations of shard archive type and platform result in a no-op.

BinaryBuilder.update_linkage — Method

update_linkage(prefix::Prefix, platform::Platform, path::AbstractString,
               old_libpath, new_libpath; verbose::Bool = false)

Given a binary object located at path within prefix, update its dynamic linkage to point to new_libpath instead of old_libpath. This is done using a tool within the cross-compilation environment such as install_name_tool on MacOS or patchelf on Linux. Windows platforms are completely skipped, as they do not encode paths or RPaths within their executables.

BinaryBuilder.versioninfo — Method

versioninfo()

Helper function to print out some debugging information

BinaryBuilder.yggdrasil_build_tarballs_path — Method

yggdrasil_build_tarballs_path(name::String)

Return the relative path within an Yggdrasil clone where this project (given its name) would be stored. This is useful for things like generating the build_tarballs.jl file and checking to see if it already exists, etc...

Note that we do not allow case-ambiguities within Yggdrasil, we check for this using the utility function case_insensitive_file_exists(path).

BinaryBuilder.yggdrasil_deploy — Method

yggdrasil_deploy(state::WizardState)

Write out a WizardState to a build_tarballs.jl in an Yggdrasil clone, then open a pull request against Yggdrasil.

BinaryBuilder.yn_prompt — Function

yn_prompt(state::WizardState, question::AbstractString, default = :y)

Perform a [Y/n] or [y/N] question loop, using default to choose between the prompt styles, and looping until a proper response (e.g. "y", "yes", "n" or "no") is received.

Pkg.PlatformEngines.package — Method

package(prefix::Prefix, output_base::AbstractString,
        version::VersionNumber;
        platform::Platform = platform_key_abi(),
        verbose::Bool = false, force::Bool = false)

Build a tarball of the prefix, storing the tarball at output_base, appending a version number, a platform-dependent suffix and a file extension. If no platform is given, defaults to current platform. Runs an audit() on the prefix, to ensure that libraries can be dlopen()'ed, that all dependencies are located within the prefix, etc... See the audit() documentation for a full list of the audit steps. Returns the full path to and hash of the generated tarball.